+++ bugsnag/src/main/java/com/bugsnag/Bugsnag.java
@@ -228,6 +228,20 @@ public class Bugsnag implements Closeable {
     }
 
     /**
+     * DEPRICATED
+     * Set which keys should be filtered when sending metaData to Bugsnag.
+     * Use this when you want to ensure sensitive information, such as passwords
+     * or credit card information is stripped from metaData you send to Bugsnag.
+     * Any keys in metaData which contain these strings will be marked as
+     * [FILTERED] when send to Bugsnag.
+     *
+     * @param filters a list of String keys to refilter from metaData
+     */
+    public void setFilters(String... filters) {
+        config.filters = filters;
+    }
+
+    /**
      * Set which keys should be redacted when sending metaData to Bugsnag.
      * Use this when you want to ensure sensitive information, such as passwords
      * or credit card information is stripped from metaData you send to Bugsnag.