+++ bugsnag/src/main/java/com/bugsnag/Report.java
@@ -3,6 +3,7 @@ package com.bugsnag;
 import com.bugsnag.serialization.Expose;
 
 import com.bugsnag.util.RedactedMap;
+import com.bugsnag.util.FilteredMap;
 
 import java.util.ArrayList;
 import java.util.Arrays;