+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 * Update `BugsnagImportSelector` to allow major versions that do not have a minor version.
   fixes [issue #211](https://github.com/bugsnag/bugsnag-java/issues/211).
-  [#]()
+  [#213](https://github.com/bugsnag/bugsnag-java/pull/213)
+
+* Add a null check for `Severity` to the notify override method. [#214](https://github.com/bugsnag/bugsnag-java/pull/214)
 
 
 ## 3.7.1 (2023-10-25)